Multiple updates
Removed intermediate variable before return, which a comment said was
necessary, but Clippy flagged it and removing it works fine now.

Replaced explicit use of From with .into() in a couple of places.

Simplified creation of a Coordinate using new From impl from a tuple.

Fixed use of tabs in a SELECT statement.