Corrected handling of stargate type
All stargates were being returned as regionals, i.e. what should have
been constellation and local jumps were all seen as regional jumps.

The pattern matching approach for comparing regions and constellations
was nice, but it wasn't clear that there was an error or why. This fix
uses explicit intermediate variables and an `if` chain to make both the
fields and the comparisons explicit.